componentize
