+++ src/routes/index.tsx
@@ -1,10 +1,10 @@
-// src/routes/index.tsx
 import { createFileRoute, useNavigate } from "@tanstack/react-router";
-import { useQuery } from "@rocicorp/zero/react";
 import { useSession } from "@/client/auth.js";
 import { queries } from "@/shared/queries.js";
-import { useZero } from "@/ui/use-zero.js";
+import { useQuery } from "@/ui/use-zero.js";
 import { useState, useEffect } from "react";
+import { Sidebar } from "@/ui/sidebar.js";
+import { MainContent } from "@/ui/main-content.js";
 
 export const Route = createFileRoute("/")({
   component: Home,
@@ -12,58 +12,38 @@ export const Route = createFileRoute("/")({
 
 function Home() {
   const { data: session } = useSession();
+  const navigate = useNavigate();
+
+  // Redirect to login if not authenticated
+  useEffect(() => {
+    if (session === null) {
+      navigate({ to: "/login" });
+    }
+  }, [session, navigate]);
   const [userTeams] = useQuery(queries.userTeams(session));
   const [selectedTeamId, setSelectedTeamId] = useState<string>("");
   const [currentView, setCurrentView] = useState<"all" | "myIssues" | "createdByMe" | "backlog" | "active" | "projects">("all");
   const [selectedIssueId, setSelectedIssueId] = useState<string | null>(null);
-  
-  // Update selected team when teams change
+
+  // Set the first team as selected when teams load
   useEffect(() => {
     if (userTeams && userTeams.length > 0 && !selectedTeamId) {
       setSelectedTeamId(userTeams[0].id);
     }
   }, [userTeams, selectedTeamId]);
-  const [showCreateTeam, setShowCreateTeam] = useState<boolean>(false);
-  const navigate = useNavigate();
 
-  if (!session) {
+  // Show loading if no teams are available yet
+  if (!userTeams || userTeams.length === 0 || !selectedTeamId) {
     return (
-      <div className="flex h-screen items-center justify-center">
+      <div className="flex h-screen bg-neutral-50 items-center justify-center">
         <div className="text-center">
-          <h1 className="text-2xl font-bold mb-4">Welcome to Linear Clone</h1>
-          <p className="text-body mb-6">Please sign in to access your issues and projects.</p>
-          <button 
-            className="btn btn-primary"
-            onClick={() => navigate({ to: "/login" })}
-          >
-            Sign In
-          </button>
-        </div>
-      </div>
-    );
-  }
-
-  if (!userTeams || userTeams.length === 0) {
-    return (
-      <div className="flex h-screen items-center justify-center">
-        <div className="text-center">
-          <h1 className="text-2xl font-bold mb-4">No Teams Found</h1>
-          <p className="text-body mb-6">You need to be part of a team to start using Linear Clone.</p>
-          <button 
-            className="btn btn-primary"
-            onClick={() => setShowCreateTeam(true)}
-          >
-            Create Team
-          </button>
-          {showCreateTeam && (
-            <CreateTeamModal 
-              onClose={() => setShowCreateTeam(false)} 
-              onSuccess={(teamId: string) => {
-                setShowCreateTeam(false);
-                setSelectedTeamId(teamId);
-                // Teams will automatically refresh due to Zero reactivity
-              }}
-            />
+          <div className="text-lg font-medium text-neutral-900 mb-2">
+            {!userTeams ? "Loading..." : "No teams found"}
+          </div>
+          {userTeams && userTeams.length === 0 && (
+            <p className="text-neutral-600">
+              Create your first team to get started with issue tracking.
+            </p>
           )}
         </div>
       </div>
@@ -75,9 +55,13 @@ function Home() {
       <Sidebar 
         teams={userTeams} 
         selectedTeamId={selectedTeamId}
-        onTeamSelect={setSelectedTeamId}
+        onTeamSelect={(teamId) => {
+          setSelectedIssueId(null); // Close any open issue detail when changing teams
+          setSelectedTeamId(teamId);
+        }}
         currentView={currentView}
         onViewChange={setCurrentView}
+        onCloseIssueDetail={() => setSelectedIssueId(null)}
       />
       <MainContent 
         teamId={selectedTeamId} 
@@ -85,1442 +69,6 @@ function Home() {
         selectedIssueId={selectedIssueId}
         onSelectIssue={setSelectedIssueId}
       />
-    </div>
-  );
-}
-
-interface SidebarProps {
-  teams: any[];
-  selectedTeamId: string;
-  onTeamSelect: (teamId: string) => void;
-  currentView: "all" | "myIssues" | "createdByMe" | "backlog" | "active" | "projects";
-  onViewChange: (view: "all" | "myIssues" | "createdByMe" | "backlog" | "active" | "projects") => void;
-}
-
-function Sidebar({ teams, selectedTeamId, onTeamSelect, currentView, onViewChange }: SidebarProps) {
-  const { data: session } = useSession();
-  const [teamIssues] = useQuery(queries.teamIssues(session, selectedTeamId, null, null, null));
-  const [userAssignedIssues] = useQuery(queries.userAssignedIssues(session, selectedTeamId));
-  
-  const openIssues = teamIssues?.filter(issue => 
-    issue.status?.type === "backlog" || 
-    issue.status?.type === "unstarted" || 
-    issue.status?.type === "started"
-  ) || [];
-  
-  const myOpenIssues = userAssignedIssues?.filter(issue =>
-    issue.status?.type === "backlog" || 
-    issue.status?.type === "unstarted" || 
-    issue.status?.type === "started"
-  ) || [];
-
-  return (
-    <aside className="w-64 bg-white border-r border-neutral-200 flex flex-col">
-      {/* Team Selector */}
-      <div className="p-4 border-b border-neutral-200">
-        <select 
-          className="input w-full"
-          value={selectedTeamId}
-          onChange={(e) => onTeamSelect(e.target.value)}
-        >
-          {teams.map(team => (
-            <option key={team.id} value={team.id}>
-              {team.name}
-            </option>
-          ))}
-        </select>
-      </div>
-
-      {/* Navigation */}
-      <nav className="flex-1 p-4">
-        <div className="space-y-2">
-          <div className="text-xs font-medium text-neutral-500 uppercase tracking-wide mb-2">
-            Your Work
-          </div>
-          
-          <button 
-            className={`nav-btn w-full text-left flex items-center justify-between ${
-              currentView === "myIssues" ? "bg-neutral-100" : ""
-            }`}
-            onClick={() => onViewChange("myIssues")}
-          >
-            <span>My Issues</span>
-            <span className="text-xs text-neutral-500">{myOpenIssues.length}</span>
-          </button>
-          
-          <button 
-            className={`nav-btn w-full text-left ${
-              currentView === "createdByMe" ? "bg-neutral-100" : ""
-            }`}
-            onClick={() => onViewChange("createdByMe")}
-          >
-            Created by me
-          </button>
-          
-          <div className="text-xs font-medium text-neutral-500 uppercase tracking-wide mb-2 mt-6">
-            Team
-          </div>
-          
-          <button 
-            className={`nav-btn w-full text-left flex items-center justify-between ${
-              currentView === "all" ? "bg-neutral-100" : ""
-            }`}
-            onClick={() => onViewChange("all")}
-          >
-            <span>All Issues</span>
-            <span className="text-xs text-neutral-500">{openIssues.length}</span>
-          </button>
-          
-          <button 
-            className={`nav-btn w-full text-left ${
-              currentView === "backlog" ? "bg-neutral-100" : ""
-            }`}
-            onClick={() => onViewChange("backlog")}
-          >
-            Backlog
-          </button>
-          
-          <button 
-            className={`nav-btn w-full text-left ${
-              currentView === "active" ? "bg-neutral-100" : ""
-            }`}
-            onClick={() => onViewChange("active")}
-          >
-            Active
-          </button>
-          
-          <button 
-            className={`nav-btn w-full text-left ${
-              currentView === "projects" ? "bg-neutral-100" : ""
-            }`}
-            onClick={() => onViewChange("projects")}
-          >
-            Projects
-          </button>
-          
-          <div className="text-xs font-medium text-neutral-500 uppercase tracking-wide mb-2 mt-6">
-            Views
-          </div>
-          
-          <button className="nav-btn w-full text-left">
-            Board
-          </button>
-          
-          <button className="nav-btn w-full text-left">
-            Timeline
-          </button>
-        </div>
-      </nav>
-
-      {/* User Menu */}
-      <div className="p-4 border-t border-neutral-200">
-        <div className="flex items-center space-x-3">
-          {session?.user.image && (
-            <img 
-              src={session.user.image} 
-              alt={session.user.name || "User"} 
-              className="w-8 h-8 rounded-full"
-            />
-          )}
-          <div className="flex-1 min-w-0">
-            <div className="text-sm font-medium text-neutral-900 truncate">
-              {session?.user.name}
-            </div>
-            <div className="text-xs text-neutral-500 truncate">
-              {session?.user.email}
-            </div>
-          </div>
-        </div>
-      </div>
-    </aside>
-  );
-}
-
-interface MainContentProps {
-  teamId: string;
-  currentView: "all" | "myIssues" | "createdByMe" | "backlog" | "active" | "projects";
-  selectedIssueId: string | null;
-  onSelectIssue: (issueId: string | null) => void;
-}
-
-function MainContent({ teamId, currentView, selectedIssueId, onSelectIssue }: MainContentProps) {
-  const { data: session } = useSession();
-  const [teamIssues] = useQuery(queries.teamIssues(session, teamId, null, null, null));
-  const [userAssignedIssues] = useQuery(queries.userAssignedIssues(session, teamId));
-  const [userCreatedIssues] = useQuery(queries.userCreatedIssues(session, teamId));
-  const [teamStatuses] = useQuery(queries.teamIssueStatuses(session, teamId));
-  const [teamPriorities] = useQuery(queries.teamIssuePriorities(session, teamId));
-  const [teamProjects] = useQuery(queries.teamProjects(session, teamId));
-  const [selectedIssue] = useQuery(selectedIssueId ? queries.issueById(session, selectedIssueId) : queries.issueById(null, ""));
-  const [displayView, setDisplayView] = useState<"list" | "board">("list");
-  const [showCreateIssue, setShowCreateIssue] = useState<boolean>(false);
-
-  // Get issues based on current view
-  const getIssuesForCurrentView = (): any[] => {
-    switch (currentView) {
-      case "myIssues":
-        return userAssignedIssues?.filter(issue => 
-          issue.status?.type === "backlog" || 
-          issue.status?.type === "unstarted" || 
-          issue.status?.type === "started"
-        ) || [];
-      case "createdByMe":
-        return userCreatedIssues || [];
-      case "backlog":
-        return teamIssues?.filter(issue => issue.status?.type === "backlog") || [];
-      case "active":
-        return teamIssues?.filter(issue => 
-          issue.status?.type === "unstarted" || 
-          issue.status?.type === "started"
-        ) || [];
-      case "projects":
-        return []; // We'll handle projects separately
-      case "all":
-      default:
-        return teamIssues || [];
-    }
-  };
-
-  const currentIssues: any[] = getIssuesForCurrentView();
-  
-  // Get view title
-  const getViewTitle = (): string => {
-    switch (currentView) {
-      case "myIssues": return "My Issues";
-      case "createdByMe": return "Created by me";
-      case "backlog": return "Backlog";
-      case "active": return "Active Issues";
-      case "projects": return "Projects";
-      case "all":
-      default: return "All Issues";
-    }
-  };
-
-  return (
-    <main className="flex-1 flex flex-col overflow-hidden">
-      {/* Header */}
-      <header className="bg-white border-b border-neutral-200 px-6 py-4">
-        <div className="flex items-center justify-between">
-          <div className="flex items-center space-x-4">
-            <h1 className="text-xl font-semibold text-neutral-900">
-              {getViewTitle()}
-            </h1>
-            {currentView !== "projects" && (
-              <div className="flex items-center space-x-2">
-                <button 
-                  className={`px-3 py-1 text-sm rounded-md ${
-                    displayView === "list" 
-                      ? "bg-neutral-100 text-neutral-900" 
-                      : "text-neutral-600 hover:text-neutral-900"
-                  }`}
-                  onClick={() => setDisplayView("list")}
-                >
-                  List
-                </button>
-                <button 
-                  className={`px-3 py-1 text-sm rounded-md ${
-                    displayView === "board" 
-                      ? "bg-neutral-100 text-neutral-900" 
-                      : "text-neutral-600 hover:text-neutral-900"
-                  }`}
-                  onClick={() => setDisplayView("board")}
-                >
-                  Board
-                </button>
-              </div>
-            )}
-          </div>
-          
-          <div className="flex items-center space-x-3">
-            {selectedIssueId && (
-              <button 
-                className="btn btn-default"
-                onClick={() => onSelectIssue(null)}
-              >
-                ← Back to {getViewTitle()}
-              </button>
-            )}
-            <button className="btn btn-default">
-              Filter
-            </button>
-            {currentView !== "projects" && (
-              <button 
-                className="btn btn-primary"
-                onClick={() => setShowCreateIssue(true)}
-              >
-                New Issue
-              </button>
-            )}
-          </div>
-        </div>
-      </header>
-
-      {/* Content */}
-      <div className="flex-1 overflow-auto">
-        {selectedIssueId && selectedIssue && selectedIssue.id ? (
-          <IssueDetail 
-            issue={selectedIssue}
-            onClose={() => onSelectIssue(null)}
-          />
-        ) : currentView === "projects" ? (
-          <ProjectsList projects={teamProjects || []} />
-        ) : displayView === "list" ? (
-          <IssuesList 
-            issues={currentIssues} 
-            onCreateIssue={() => setShowCreateIssue(true)}
-            onSelectIssue={onSelectIssue}
-          />
-        ) : (
-          <IssuesBoard 
-            issues={currentIssues} 
-            statuses={teamStatuses || []}
-            onSelectIssue={onSelectIssue}
-          />
-        )}
-      </div>
-      
-      {showCreateIssue && (
-        <CreateIssueModal 
-          teamId={teamId}
-          statuses={teamStatuses || []}
-          priorities={teamPriorities || []}
-          onClose={() => setShowCreateIssue(false)} 
-          onSuccess={() => setShowCreateIssue(false)}
-        />
-      )}
-    </main>
-  );
-}
-
-interface IssuesListProps {
-  issues: any[];
-  onCreateIssue?: () => void;
-  onSelectIssue?: (issueId: string) => void;
-}
-
-function IssuesList({ issues, onCreateIssue, onSelectIssue }: IssuesListProps) {
-  if (issues.length === 0) {
-    return (
-      <div className="p-8 text-center">
-        <div className="text-neutral-500 mb-4">No issues found</div>
-        <button 
-          className="btn btn-primary"
-          onClick={onCreateIssue}
-        >
-          Create your first issue
-        </button>
-      </div>
-    );
-  }
-
-  return (
-    <div className="p-6">
-      <div className="space-y-1">
-        {issues.map((issue) => (
-          <div 
-            key={issue.id}
-            className="flex items-center px-4 py-3 bg-white rounded-lg border border-neutral-200 hover:border-neutral-300 cursor-pointer transition-colors"
-            onClick={() => onSelectIssue?.(issue.id)}
-          >
-            <div className="flex items-center space-x-3 flex-1">
-              <div className="flex items-center space-x-2">
-                <div className={`w-2 h-2 rounded-full ${
-                  issue.priority?.value >= 3 ? "bg-red-500" :
-                  issue.priority?.value >= 2 ? "bg-orange-500" :
-                  issue.priority?.value >= 1 ? "bg-yellow-500" :
-                  "bg-gray-400"
-                }`} />
-                <span className="text-sm text-neutral-500">
-                  {issue.team?.identifier}-{issue.number}
-                </span>
-              </div>
-              
-              <div className="flex-1">
-                <div className="font-medium text-neutral-900">
-                  {issue.title}
-                </div>
-              </div>
-              
-              <div className="flex items-center space-x-3">
-                <div className={`px-2 py-1 text-xs rounded-full ${
-                  issue.status?.type === "completed" ? "bg-green-100 text-green-800" :
-                  issue.status?.type === "started" ? "bg-blue-100 text-blue-800" :
-                  issue.status?.type === "unstarted" ? "bg-yellow-100 text-yellow-800" :
-                  "bg-gray-100 text-gray-800"
-                }`}>
-                  {issue.status?.name}
-                </div>
-                
-                {issue.assignee && (
-                  <div className="flex items-center space-x-2">
-                    {issue.assignee.image && (
-                      <img 
-                        src={issue.assignee.image} 
-                        alt={issue.assignee.name} 
-                        className="w-6 h-6 rounded-full"
-                      />
-                    )}
-                    <span className="text-sm text-neutral-600">
-                      {issue.assignee.name}
-                    </span>
-                  </div>
-                )}
-              </div>
-            </div>
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-}
-
-interface IssuesBoardProps {
-  issues: any[];
-  statuses: any[];
-  onSelectIssue?: (issueId: string) => void;
-}
-
-function IssuesBoard({ issues, statuses, onSelectIssue }: IssuesBoardProps) {
-  const groupedIssues = statuses.reduce((acc, status) => {
-    acc[status.id] = issues.filter(issue => issue.statusId === status.id);
-    return acc;
-  }, {} as Record<string, any[]>);
-
-  return (
-    <div className="flex gap-6 p-6 overflow-x-auto min-h-full">
-      {statuses.map((status) => (
-        <div key={status.id} className="flex-shrink-0 w-80">
-          <div className="mb-4">
-            <h3 className="font-medium text-neutral-900 mb-1">
-              {status.name}
-            </h3>
-            <div className="text-sm text-neutral-500">
-              {groupedIssues[status.id]?.length || 0} issues
-            </div>
-          </div>
-          
-          <div className="space-y-3">
-            {(groupedIssues[status.id] || []).map((issue) => (
-              <div 
-                key={issue.id}
-                className="bg-white p-4 rounded-lg border border-neutral-200 hover:border-neutral-300 cursor-pointer transition-colors"
-                onClick={() => onSelectIssue?.(issue.id)}
-              >
-                <div className="mb-2">
-                  <div className="flex items-center space-x-2 mb-2">
-                    <div className={`w-2 h-2 rounded-full ${
-                      issue.priority?.value >= 3 ? "bg-red-500" :
-                      issue.priority?.value >= 2 ? "bg-orange-500" :
-                      issue.priority?.value >= 1 ? "bg-yellow-500" :
-                      "bg-gray-400"
-                    }`} />
-                    <span className="text-xs text-neutral-500">
-                      {issue.team?.identifier}-{issue.number}
-                    </span>
-                  </div>
-                  
-                  <div className="font-medium text-neutral-900 text-sm">
-                    {issue.title}
-                  </div>
-                </div>
-                
-                {issue.assignee && (
-                  <div className="flex items-center space-x-2">
-                    {issue.assignee.image && (
-                      <img 
-                        src={issue.assignee.image} 
-                        alt={issue.assignee.name} 
-                        className="w-5 h-5 rounded-full"
-                      />
-                    )}
-                    <span className="text-xs text-neutral-600">
-                      {issue.assignee.name}
-                    </span>
-                  </div>
-                )}
-              </div>
-            ))}
-          </div>
-        </div>
-      ))}
-    </div>
-  );
-}
-
-interface IssueDetailProps {
-  issue: any;
-  onClose: () => void;
-}
-
-function IssueDetail({ issue, onClose }: IssueDetailProps) {
-  const zero = useZero();
-  const { data: session } = useSession();
-  const [teamMembers] = useQuery(queries.teamMembers(session, issue.teamId));
-  const [showEditModal, setShowEditModal] = useState<boolean>(false);
-  const [commentText, setCommentText] = useState<string>("");
-  const [isEditingTitle, setIsEditingTitle] = useState<boolean>(false);
-  const [isEditingDescription, setIsEditingDescription] = useState<boolean>(false);
-  const [editTitle, setEditTitle] = useState<string>(issue.title || "");
-  const [editDescription, setEditDescription] = useState<string>(issue.description || "");
-
-  // Check if user can edit this issue
-  const canEdit: boolean = session ? (
-    session.user.id === issue.creatorId || 
-    session.user.id === issue.assigneeId ||
-    teamMembers?.some(member => 
-      member.userId === session.user.id && member.role === "admin"
-    )
-  ) : false;
-
-  const handleAddComment = async (): Promise<void> => {
-    if (!commentText.trim()) return;
-    
-    try {
-      const commentId: string = `comment_${Date.now()}_${Math.random().toString(36).substring(2)}`;
-      await zero.mutate.createIssueComment({
-        id: commentId,
-        issueId: issue.id,
-        body: commentText.trim(),
-      });
-      setCommentText("");
-    } catch (err) {
-      console.error("Failed to add comment:", err);
-    }
-  };
-
-  const handleDeleteIssue = async (): Promise<void> => {
-    if (!confirm(`Are you sure you want to delete issue ${issue.team?.identifier}-${issue.number}? This action cannot be undone.`)) {
-      return;
-    }
-    
-    try {
-      await zero.mutate.deleteIssue({
-        id: issue.id,
-      });
-      onClose(); // Close the detail view after deletion
-    } catch (err) {
-      console.error("Failed to delete issue:", err);
-      alert("Failed to delete issue. Please try again.");
-    }
-  };
-
-  const handleSaveTitle = async (): Promise<void> => {
-    if (!editTitle.trim()) return;
-    
-    try {
-      await zero.mutate.updateIssue({
-        id: issue.id,
-        title: editTitle.trim(),
-      });
-      setIsEditingTitle(false);
-    } catch (err) {
-      console.error("Failed to update title:", err);
-      setEditTitle(issue.title || ""); // Reset on error
-    }
-  };
-
-  const handleSaveDescription = async (): Promise<void> => {
-    try {
-      await zero.mutate.updateIssue({
-        id: issue.id,
-        description: editDescription.trim() || undefined,
-      });
-      setIsEditingDescription(false);
-    } catch (err) {
-      console.error("Failed to update description:", err);
-      setEditDescription(issue.description || ""); // Reset on error
-    }
-  };
-
-  const handleCancelTitleEdit = (): void => {
-    setEditTitle(issue.title || "");
-    setIsEditingTitle(false);
-  };
-
-  const handleCancelDescriptionEdit = (): void => {
-    setEditDescription(issue.description || "");
-    setIsEditingDescription(false);
-  };
-
-  return (
-    <div className="flex h-full">
-      {/* Main content */}
-      <div className="flex-1 p-6">
-        <div className="max-w-4xl mx-auto">
-          {/* Issue header */}
-          <div className="mb-6">
-            <div className="flex items-center justify-between mb-4">
-              <div className="flex items-center space-x-2">
-                <div className={`w-3 h-3 rounded-full ${
-                  issue.priority?.value >= 3 ? "bg-red-500" :
-                  issue.priority?.value >= 2 ? "bg-orange-500" :
-                  issue.priority?.value >= 1 ? "bg-yellow-500" :
-                  "bg-gray-400"
-                }`} />
-                <span className="text-sm text-neutral-500">
-                  {issue.team?.identifier}-{issue.number}
-                </span>
-                <div className={`px-2 py-1 text-xs rounded-full ${
-                  issue.status?.type === "completed" ? "bg-green-100 text-green-800" :
-                  issue.status?.type === "started" ? "bg-blue-100 text-blue-800" :
-                  issue.status?.type === "unstarted" ? "bg-yellow-100 text-yellow-800" :
-                  "bg-gray-100 text-gray-800"
-                }`}>
-                  {issue.status?.name}
-                </div>
-              </div>
-              
-              {/* Edit and Delete buttons - only show if user can edit */}
-              {canEdit && (
-                <div className="flex space-x-2">
-                  <button 
-                    className="btn btn-default"
-                    onClick={() => setShowEditModal(true)}
-                  >
-                    Edit Issue
-                  </button>
-                  <button 
-                    className="btn btn-default text-red-600 hover:bg-red-50"
-                    onClick={() => handleDeleteIssue()}
-                  >
-                    Delete
-                  </button>
-                </div>
-              )}
-            </div>
-            
-            {/* Title - inline editing */}
-            {isEditingTitle ? (
-              <div className="mb-4">
-                <input
-                  type="text"
-                  value={editTitle}
-                  onChange={(e) => setEditTitle(e.target.value)}
-                  className="text-2xl font-bold text-neutral-900 w-full border-none outline-none bg-transparent border-b-2 border-blue-500 focus:border-blue-600"
-                  onBlur={handleSaveTitle}
-                  onKeyDown={(e) => {
-                    if (e.key === 'Enter') {
-                      handleSaveTitle();
-                    } else if (e.key === 'Escape') {
-                      handleCancelTitleEdit();
-                    }
-                  }}
-                  autoFocus
-                />
-                <div className="flex space-x-2 mt-2">
-                  <button
-                    onClick={handleSaveTitle}
-                    className="text-xs bg-blue-600 text-white px-2 py-1 rounded hover:bg-blue-700"
-                  >
-                    Save
-                  </button>
-                  <button
-                    onClick={handleCancelTitleEdit}
-                    className="text-xs bg-gray-500 text-white px-2 py-1 rounded hover:bg-gray-600"
-                  >
-                    Cancel
-                  </button>
-                </div>
-              </div>
-            ) : (
-              <h1 
-                className={`text-2xl font-bold text-neutral-900 mb-4 ${canEdit ? 'cursor-pointer hover:bg-gray-50 rounded px-1' : ''}`}
-                onClick={canEdit ? () => setIsEditingTitle(true) : undefined}
-                title={canEdit ? "Click to edit title" : undefined}
-              >
-                {issue.title}
-              </h1>
-            )}
-            
-            {/* Description - inline editing */}
-            {isEditingDescription ? (
-              <div className="mb-6">
-                <textarea
-                  value={editDescription}
-                  onChange={(e) => setEditDescription(e.target.value)}
-                  className="text-neutral-700 w-full min-h-[100px] border-2 border-blue-500 rounded p-2 focus:border-blue-600 outline-none resize-vertical"
-                  onBlur={handleSaveDescription}
-                  onKeyDown={(e) => {
-                    if (e.key === 'Escape') {
-                      handleCancelDescriptionEdit();
-                    }
-                    // Allow Ctrl+Enter or Cmd+Enter to save
-                    if ((e.ctrlKey || e.metaKey) && e.key === 'Enter') {
-                      handleSaveDescription();
-                    }
-                  }}
-                  placeholder="Add a description..."
-                  autoFocus
-                />
-                <div className="flex space-x-2 mt-2">
-                  <button
-                    onClick={handleSaveDescription}
-                    className="text-xs bg-blue-600 text-white px-2 py-1 rounded hover:bg-blue-700"
-                  >
-                    Save
-                  </button>
-                  <button
-                    onClick={handleCancelDescriptionEdit}
-                    className="text-xs bg-gray-500 text-white px-2 py-1 rounded hover:bg-gray-600"
-                  >
-                    Cancel
-                  </button>
-                </div>
-              </div>
-            ) : (
-              <div 
-                className={`text-neutral-700 mb-6 whitespace-pre-wrap min-h-[40px] ${canEdit ? 'cursor-pointer hover:bg-gray-50 rounded p-1' : ''}`}
-                onClick={canEdit ? () => setIsEditingDescription(true) : undefined}
-                title={canEdit ? "Click to edit description" : undefined}
-              >
-                {issue.description || (canEdit ? <span className="text-gray-400 italic">Click to add description...</span> : null)}
-              </div>
-            )}
-          </div>
-
-          {/* Issue metadata */}
-          <div className="grid grid-cols-2 gap-6 mb-8">
-            <div>
-              <h3 className="text-sm font-medium text-neutral-500 mb-2">Assignee</h3>
-              <div className="flex items-center space-x-2">
-                {issue.assignee?.image && (
-                  <img 
-                    src={issue.assignee.image} 
-                    alt={issue.assignee.name} 
-                    className="w-6 h-6 rounded-full"
-                  />
-                )}
-                <span className="text-sm">
-                  {issue.assignee?.name || "Unassigned"}
-                </span>
-              </div>
-            </div>
-            
-            <div>
-              <h3 className="text-sm font-medium text-neutral-500 mb-2">Priority</h3>
-              <span className="text-sm">{issue.priority?.name || "No Priority"}</span>
-            </div>
-            
-            <div>
-              <h3 className="text-sm font-medium text-neutral-500 mb-2">Created by</h3>
-              <div className="flex items-center space-x-2">
-                {issue.creator?.image && (
-                  <img 
-                    src={issue.creator.image} 
-                    alt={issue.creator.name} 
-                    className="w-6 h-6 rounded-full"
-                  />
-                )}
-                <span className="text-sm">{issue.creator?.name}</span>
-              </div>
-            </div>
-            
-            <div>
-              <h3 className="text-sm font-medium text-neutral-500 mb-2">Project</h3>
-              <span className="text-sm">{issue.project?.name || "No Project"}</span>
-            </div>
-          </div>
-
-          {/* Comments section */}
-          <div className="border-t border-neutral-200 pt-6">
-            <h3 className="text-lg font-semibold mb-4">Activity</h3>
-            
-            {/* Comment input */}
-            <div className="mb-6">
-              <textarea
-                className="input w-full"
-                rows={3}
-                value={commentText}
-                onChange={(e) => setCommentText(e.target.value)}
-                placeholder="Add a comment..."
-              />
-              <div className="flex justify-end mt-2">
-                <button
-                  className="btn btn-primary"
-                  onClick={handleAddComment}
-                  disabled={!commentText.trim()}
-                >
-                  Comment
-                </button>
-              </div>
-            </div>
-
-            {/* Comments list */}
-            <div className="space-y-4">
-              {issue.comments?.map((comment: any) => (
-                <div key={comment.id} className="bg-neutral-50 rounded-lg p-4">
-                  <div className="flex items-center space-x-2 mb-2">
-                    {comment.author?.image && (
-                      <img 
-                        src={comment.author.image} 
-                        alt={comment.author.name} 
-                        className="w-6 h-6 rounded-full"
-                      />
-                    )}
-                    <span className="font-medium text-sm">{comment.author?.name}</span>
-                    <span className="text-xs text-neutral-500">
-                      {new Date(comment.createdAt).toLocaleDateString()}
-                    </span>
-                  </div>
-                  <div className="text-sm text-neutral-700 whitespace-pre-wrap">
-                    {comment.body}
-                  </div>
-                </div>
-              ))}
-            </div>
-          </div>
-        </div>
-      </div>
-      
-      {/* Edit Modal */}
-      {showEditModal && (
-        <EditIssueModal 
-          issue={issue}
-          onClose={() => setShowEditModal(false)} 
-          onSuccess={() => setShowEditModal(false)}
-        />
-      )}
-    </div>
-  );
-}
-
-interface ProjectsListProps {
-  projects: any[];
-}
-
-function ProjectsList({ projects }: ProjectsListProps) {
-  if (projects.length === 0) {
-    return (
-      <div className="p-8 text-center">
-        <div className="text-neutral-500 mb-4">No projects found</div>
-        <button className="btn btn-primary">
-          Create your first project
-        </button>
-      </div>
-    );
-  }
-
-  return (
-    <div className="p-6">
-      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-        {projects.map((project) => (
-          <div 
-            key={project.id}
-            className="bg-white p-6 rounded-lg border border-neutral-200 hover:border-neutral-300 cursor-pointer transition-colors"
-          >
-            <h3 className="font-semibold text-lg mb-2">{project.name}</h3>
-            {project.description && (
-              <p className="text-neutral-600 text-sm mb-4 line-clamp-3">
-                {project.description}
-              </p>
-            )}
-            <div className="flex items-center justify-between text-sm text-neutral-500">
-              <span>{project.identifier}</span>
-              {project.lead && (
-                <div className="flex items-center space-x-1">
-                  {project.lead.image && (
-                    <img 
-                      src={project.lead.image} 
-                      alt={project.lead.name} 
-                      className="w-4 h-4 rounded-full"
-                    />
-                  )}
-                  <span>{project.lead.name}</span>
-                </div>
-              )}
-            </div>
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-}
-
-interface EditIssueModalProps {
-  issue: any;
-  onClose: () => void;
-  onSuccess: () => void;
-}
-
-function EditIssueModal({ issue, onClose, onSuccess }: EditIssueModalProps) {
-  const zero = useZero();
-  const { data: session } = useSession();
-  const [teamMembers] = useQuery(queries.teamMembers(session, issue.teamId));
-  const [teamStatuses] = useQuery(queries.teamIssueStatuses(session, issue.teamId));
-  const [teamPriorities] = useQuery(queries.teamIssuePriorities(session, issue.teamId));
-  
-  const [formData, setFormData] = useState({
-    title: issue.title || "",
-    description: issue.description || "",
-    statusId: issue.statusId || "",
-    priorityId: issue.priorityId || "",
-    assigneeId: issue.assigneeId || "",
-    estimate: issue.estimate?.toString() || "",
-  });
-  const [isSubmitting, setIsSubmitting] = useState<boolean>(false);
-  const [error, setError] = useState<string>("");
-
-  const handleSubmit = async (e: React.FormEvent): Promise<void> => {
-    e.preventDefault();
-    
-    if (!formData.title.trim()) {
-      setError("Issue title is required");
-      return;
-    }
-
-    setIsSubmitting(true);
-    setError("");
-
-    try {
-      await zero.mutate.updateIssue({
-        id: issue.id,
-        title: formData.title.trim(),
-        description: formData.description.trim() || undefined,
-        statusId: formData.statusId || undefined,
-        priorityId: formData.priorityId || undefined,
-        assigneeId: formData.assigneeId || undefined,
-        estimate: formData.estimate ? parseInt(formData.estimate) : undefined,
-      });
-
-      onSuccess();
-    } catch (err) {
-      console.error("Failed to update issue:", err);
-      setError(err instanceof Error ? err.message : "Failed to update issue");
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
-
-  return (
-    <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50">
-      <div className="bg-white rounded-lg p-6 w-full max-w-lg mx-4 max-h-[90vh] overflow-y-auto">
-        <div className="flex justify-between items-center mb-4">
-          <h2 className="text-xl font-semibold">Edit Issue</h2>
-          <button 
-            onClick={onClose}
-            className="text-neutral-400 hover:text-neutral-600"
-          >
-            <svg className="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-            </svg>
-          </button>
-        </div>
-
-        <form onSubmit={handleSubmit} className="space-y-4">
-          <div>
-            <label className="block text-sm font-medium text-neutral-700 mb-1">
-              Title *
-            </label>
-            <input
-              type="text"
-              className="input w-full"
-              value={formData.title}
-              onChange={(e) => setFormData(prev => ({ ...prev, title: e.target.value }))}
-              placeholder="What needs to be done?"
-              maxLength={200}
-            />
-          </div>
-
-          <div>
-            <label className="block text-sm font-medium text-neutral-700 mb-1">
-              Description
-            </label>
-            <textarea
-              className="input w-full"
-              rows={4}
-              value={formData.description}
-              onChange={(e) => setFormData(prev => ({ ...prev, description: e.target.value }))}
-              placeholder="Add more details..."
-              maxLength={1000}
-            />
-          </div>
-
-          <div className="grid grid-cols-2 gap-4">
-            <div>
-              <label className="block text-sm font-medium text-neutral-700 mb-1">
-                Status
-              </label>
-              <select
-                className="input w-full"
-                value={formData.statusId}
-                onChange={(e) => setFormData(prev => ({ ...prev, statusId: e.target.value }))}
-              >
-                {teamStatuses?.map(status => (
-                  <option key={status.id} value={status.id}>
-                    {status.name}
-                  </option>
-                ))}
-              </select>
-            </div>
-
-            <div>
-              <label className="block text-sm font-medium text-neutral-700 mb-1">
-                Priority
-              </label>
-              <select
-                className="input w-full"
-                value={formData.priorityId}
-                onChange={(e) => setFormData(prev => ({ ...prev, priorityId: e.target.value }))}
-              >
-                {teamPriorities?.map(priority => (
-                  <option key={priority.id} value={priority.id}>
-                    {priority.name}
-                  </option>
-                ))}
-              </select>
-            </div>
-          </div>
-
-          <div className="grid grid-cols-2 gap-4">
-            <div>
-              <label className="block text-sm font-medium text-neutral-700 mb-1">
-                Assignee
-              </label>
-              <select
-                className="input w-full"
-                value={formData.assigneeId}
-                onChange={(e) => setFormData(prev => ({ ...prev, assigneeId: e.target.value }))}
-              >
-                <option value="">Unassigned</option>
-                {teamMembers?.map(member => (
-                  <option key={member.userId} value={member.userId}>
-                    {member.user?.name || member.user?.email}
-                  </option>
-                ))}
-              </select>
-            </div>
-
-            <div>
-              <label className="block text-sm font-medium text-neutral-700 mb-1">
-                Estimate (hours)
-              </label>
-              <input
-                type="number"
-                className="input w-full"
-                value={formData.estimate}
-                onChange={(e) => setFormData(prev => ({ ...prev, estimate: e.target.value }))}
-                placeholder="0"
-                min="0"
-                max="999"
-              />
-            </div>
-          </div>
-
-          {error && (
-            <div className="text-red-600 text-sm">
-              {error}
-            </div>
-          )}
-
-          <div className="flex justify-end space-x-3 pt-4">
-            <button
-              type="button"
-              onClick={onClose}
-              className="btn btn-default"
-              disabled={isSubmitting}
-            >
-              Cancel
-            </button>
-            <button
-              type="submit"
-              className="btn btn-primary"
-              disabled={isSubmitting}
-            >
-              {isSubmitting ? "Updating..." : "Update Issue"}
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
-  );
-}
-
-interface CreateTeamModalProps {
-  onClose: () => void;
-  onSuccess: (teamId: string) => void;
-}
-
-function CreateTeamModal({ onClose, onSuccess }: CreateTeamModalProps) {
-  const zero = useZero();
-  const [formData, setFormData] = useState({
-    name: "",
-    identifier: "",
-    description: "",
-  });
-  const [isSubmitting, setIsSubmitting] = useState<boolean>(false);
-  const [error, setError] = useState<string>("");
-
-  const handleSubmit = async (e: React.FormEvent): Promise<void> => {
-    e.preventDefault();
-    
-    if (!formData.name.trim()) {
-      setError("Team name is required");
-      return;
-    }
-    
-    if (!formData.identifier.trim()) {
-      setError("Team identifier is required");
-      return;
-    }
-
-    // Validate identifier (should be uppercase letters and numbers only)
-    if (!/^[A-Z0-9]+$/.test(formData.identifier)) {
-      setError("Identifier should only contain uppercase letters and numbers");
-      return;
-    }
-
-    setIsSubmitting(true);
-    setError("");
-
-    try {
-      const teamId: string = `team_${Date.now()}_${Math.random().toString(36).substring(2)}`;
-      
-      await zero.mutate.createTeam({
-        id: teamId,
-        name: formData.name.trim(),
-        identifier: formData.identifier.trim().toUpperCase(),
-        description: formData.description.trim() || undefined,
-      });
-
-      // Create default statuses for the new team
-      const statusesData = [
-        { name: "Backlog", type: "backlog" as const, position: 0 },
-        { name: "Todo", type: "unstarted" as const, position: 1 },
-        { name: "In Progress", type: "started" as const, position: 2 },
-        { name: "Done", type: "completed" as const, position: 3 },
-      ];
-
-      for (const status of statusesData) {
-        const statusId: string = `status_${teamId}_${status.type}`;
-        await zero.mutate.createIssueStatus({
-          id: statusId,
-          name: status.name,
-          type: status.type,
-          position: status.position,
-          teamId,
-        });
-      }
-
-      // Create default priorities
-      const prioritiesData = [
-        { name: "No Priority", value: 0 },
-        { name: "Low", value: 1 },
-        { name: "Medium", value: 2 },
-        { name: "High", value: 3 },
-        { name: "Urgent", value: 4 },
-      ];
-
-      for (const priority of prioritiesData) {
-        const priorityId: string = `priority_${teamId}_${priority.value}`;
-        await zero.mutate.createIssuePriority({
-          id: priorityId,
-          name: priority.name,
-          value: priority.value,
-          teamId,
-        });
-      }
-
-      onSuccess(teamId);
-    } catch (err) {
-      console.error("Failed to create team:", err);
-      setError(err instanceof Error ? err.message : "Failed to create team");
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
-
-  return (
-    <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50">
-      <div className="bg-white rounded-lg p-6 w-full max-w-md mx-4">
-        <div className="flex justify-between items-center mb-4">
-          <h2 className="text-xl font-semibold">Create Team</h2>
-          <button 
-            onClick={onClose}
-            className="text-neutral-400 hover:text-neutral-600"
-          >
-            <svg className="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-            </svg>
-          </button>
-        </div>
-
-        <form onSubmit={handleSubmit} className="space-y-4">
-          <div>
-            <label className="block text-sm font-medium text-neutral-700 mb-1">
-              Team Name *
-            </label>
-            <input
-              type="text"
-              className="input w-full"
-              value={formData.name}
-              onChange={(e) => setFormData(prev => ({ ...prev, name: e.target.value }))}
-              placeholder="My Awesome Team"
-              maxLength={50}
-            />
-          </div>
-
-          <div>
-            <label className="block text-sm font-medium text-neutral-700 mb-1">
-              Identifier *
-            </label>
-            <input
-              type="text"
-              className="input w-full"
-              value={formData.identifier}
-              onChange={(e) => setFormData(prev => ({ 
-                ...prev, 
-                identifier: e.target.value.toUpperCase().replace(/[^A-Z0-9]/g, "")
-              }))}
-              placeholder="MAT"
-              maxLength={10}
-            />
-            <p className="text-xs text-neutral-500 mt-1">
-              Used for issue numbering (e.g., MAT-123). Uppercase letters and numbers only.
-            </p>
-          </div>
-
-          <div>
-            <label className="block text-sm font-medium text-neutral-700 mb-1">
-              Description
-            </label>
-            <textarea
-              className="input w-full"
-              rows={3}
-              value={formData.description}
-              onChange={(e) => setFormData(prev => ({ ...prev, description: e.target.value }))}
-              placeholder="Optional description of your team..."
-              maxLength={200}
-            />
-          </div>
-
-          {error && (
-            <div className="text-red-600 text-sm">
-              {error}
-            </div>
-          )}
-
-          <div className="flex justify-end space-x-3 pt-4">
-            <button
-              type="button"
-              onClick={onClose}
-              className="btn btn-default"
-              disabled={isSubmitting}
-            >
-              Cancel
-            </button>
-            <button
-              type="submit"
-              className="btn btn-primary"
-              disabled={isSubmitting}
-            >
-              {isSubmitting ? "Creating..." : "Create Team"}
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
-  );
-}
-
-interface CreateIssueModalProps {
-  teamId: string;
-  statuses: any[];
-  priorities: any[];
-  onClose: () => void;
-  onSuccess: () => void;
-}
-
-function CreateIssueModal({ teamId, statuses, priorities, onClose, onSuccess }: CreateIssueModalProps) {
-  const zero = useZero();
-  const { data: session } = useSession();
-  const [teamMembers] = useQuery(queries.teamMembers(session, teamId));
-  
-  const [formData, setFormData] = useState({
-    title: "",
-    description: "",
-    statusId: statuses.find(s => s.type === "backlog")?.id || statuses[0]?.id || "",
-    priorityId: priorities.find(p => p.value === 0)?.id || "",
-    assigneeId: "",
-    estimate: "",
-  });
-  const [isSubmitting, setIsSubmitting] = useState<boolean>(false);
-  const [error, setError] = useState<string>("");
-
-  const handleSubmit = async (e: React.FormEvent): Promise<void> => {
-    e.preventDefault();
-    
-    if (!formData.title.trim()) {
-      setError("Issue title is required");
-      return;
-    }
-
-    setIsSubmitting(true);
-    setError("");
-
-    try {
-      const issueId: string = `issue_${Date.now()}_${Math.random().toString(36).substring(2)}`;
-      
-      await zero.mutate.createIssue({
-        id: issueId,
-        title: formData.title.trim(),
-        description: formData.description.trim() || undefined,
-        statusId: formData.statusId,
-        priorityId: formData.priorityId || undefined,
-        projectId: undefined,
-        teamId,
-        assigneeId: formData.assigneeId || undefined,
-        estimate: formData.estimate ? parseInt(formData.estimate) : undefined,
-      });
-
-      onSuccess();
-    } catch (err) {
-      console.error("Failed to create issue:", err);
-      setError(err instanceof Error ? err.message : "Failed to create issue");
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
-
-  return (
-    <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50">
-      <div className="bg-white rounded-lg p-6 w-full max-w-lg mx-4 max-h-[90vh] overflow-y-auto">
-        <div className="flex justify-between items-center mb-4">
-          <h2 className="text-xl font-semibold">Create Issue</h2>
-          <button 
-            onClick={onClose}
-            className="text-neutral-400 hover:text-neutral-600"
-          >
-            <svg className="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-            </svg>
-          </button>
-        </div>
-
-        <form onSubmit={handleSubmit} className="space-y-4">
-          <div>
-            <label className="block text-sm font-medium text-neutral-700 mb-1">
-              Title *
-            </label>
-            <input
-              type="text"
-              className="input w-full"
-              value={formData.title}
-              onChange={(e) => setFormData(prev => ({ ...prev, title: e.target.value }))}
-              placeholder="What needs to be done?"
-              maxLength={200}
-            />
-          </div>
-
-          <div>
-            <label className="block text-sm font-medium text-neutral-700 mb-1">
-              Description
-            </label>
-            <textarea
-              className="input w-full"
-              rows={4}
-              value={formData.description}
-              onChange={(e) => setFormData(prev => ({ ...prev, description: e.target.value }))}
-              placeholder="Add more details..."
-              maxLength={1000}
-            />
-          </div>
-
-          <div className="grid grid-cols-2 gap-4">
-            <div>
-              <label className="block text-sm font-medium text-neutral-700 mb-1">
-                Status
-              </label>
-              <select
-                className="input w-full"
-                value={formData.statusId}
-                onChange={(e) => setFormData(prev => ({ ...prev, statusId: e.target.value }))}
-              >
-                {statuses.map(status => (
-                  <option key={status.id} value={status.id}>
-                    {status.name}
-                  </option>
-                ))}
-              </select>
-            </div>
-
-            <div>
-              <label className="block text-sm font-medium text-neutral-700 mb-1">
-                Priority
-              </label>
-              <select
-                className="input w-full"
-                value={formData.priorityId}
-                onChange={(e) => setFormData(prev => ({ ...prev, priorityId: e.target.value }))}
-              >
-                {priorities.map(priority => (
-                  <option key={priority.id} value={priority.id}>
-                    {priority.name}
-                  </option>
-                ))}
-              </select>
-            </div>
-          </div>
-
-          <div className="grid grid-cols-2 gap-4">
-            <div>
-              <label className="block text-sm font-medium text-neutral-700 mb-1">
-                Assignee
-              </label>
-              <select
-                className="input w-full"
-                value={formData.assigneeId}
-                onChange={(e) => setFormData(prev => ({ ...prev, assigneeId: e.target.value }))}
-              >
-                <option value="">Unassigned</option>
-                {teamMembers?.map(member => (
-                  <option key={member.userId} value={member.userId}>
-                    {member.user?.name || member.user?.email}
-                  </option>
-                ))}
-              </select>
-            </div>
-
-            <div>
-              <label className="block text-sm font-medium text-neutral-700 mb-1">
-                Estimate (hours)
-              </label>
-              <input
-                type="number"
-                className="input w-full"
-                value={formData.estimate}
-                onChange={(e) => setFormData(prev => ({ ...prev, estimate: e.target.value }))}
-                placeholder="0"
-                min="0"
-                max="999"
-              />
-            </div>
-          </div>
-
-          {error && (
-            <div className="text-red-600 text-sm">
-              {error}
-            </div>
-          )}
-
-          <div className="flex justify-end space-x-3 pt-4">
-            <button
-              type="button"
-              onClick={onClose}
-              className="btn btn-default"
-              disabled={isSubmitting}
-            >
-              Cancel
-            </button>
-            <button
-              type="submit"
-              className="btn btn-primary"
-              disabled={isSubmitting}
-            >
-              {isSubmitting ? "Creating..." : "Create Issue"}
-            </button>
-          </div>
-        </form>
-      </div>
     </div>
   );
 }
